% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-colors.R
\name{adaptive_color}
\alias{adaptive_color}
\alias{theme_background_color}
\alias{theme_foreground_color}
\title{Adaptive colors colors for {hrbragg} themes}
\usage{
adaptive_color(mode = c("light", "dark", "rstudio"), light, dark)

theme_background_color(
  mode = c("light", "dark", "rstudio"),
  light = "white",
  dark = "#121212"
)

theme_foreground_color(
  mode = c("light", "dark", "rstudio"),
  light = "#262626",
  dark = "#e1e1e1"
)
}
\arguments{
\item{mode}{\code{dark}, \code{light}, or \code{rstudio} to base it on RStudio's active theme}

\item{light, dark}{light and dark colors to auto-select from}

\item{dark}{dark color to use}
}
\value{
length 1 character vector that represents a valid color
}
\description{
Explicit foreground/background theme color functions are provided
along with the primary adaptive color switcher function.
}
\examples{
theme_background_color()
}
