% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{ell}
\alias{ell}
\title{Shortcut for \code{element_line}}
\usage{
ell(
  colour = NULL,
  size = NULL,
  linetype = NULL,
  lineend = NULL,
  color = NULL,
  arrow = NULL,
  inherit.blank = FALSE
)
}
\arguments{
\item{colour, size, linetype, lineend, color, arrow, inherit.blank}{See \code{\link[ggplot2:element]{ggplot2::element_line()}}}
}
\value{
An S3 object of class element, rel, or margin.
}
\description{
Shortcut for \code{element_line}
}
