% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{elr}
\alias{elr}
\title{Shortcut for \code{element_rect}}
\usage{
elr(
  fill = NULL,
  colour = NULL,
  size = NULL,
  linetype = NULL,
  color = NULL,
  inherit.blank = FALSE
)
}
\arguments{
\item{fill, colour, size, linetype, color, inherit.blank}{See \code{\link[ggplot2:element]{ggplot2::element_rect()}}}
}
\value{
An S3 object of class element, rel, or margin.
}
\description{
Shortcut for \code{element_rect}
}
