% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcuts.R
\name{elt}
\alias{elt}
\title{Shortcut for \code{element_text}}
\usage{
elt(
  family = NULL,
  face = NULL,
  colour = NULL,
  size = NULL,
  hjust = NULL,
  vjust = NULL,
  angle = NULL,
  lineheight = NULL,
  color = NULL,
  margin = NULL,
  debug = NULL,
  inherit.blank = FALSE
)
}
\arguments{
\item{family, face, colour, size, hjust, vjust, angle, lineheight, color, margin, debug, inherit.blank}{See \code{\link[ggplot2:element]{ggplot2::element_text()}}}
}
\description{
Shortcut for \code{element_text}
}
