% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-dict.R
\docType{data}
\name{feature_dict}
\alias{feature_dict}
\title{OpenType feature description lookup table}
\format{
A data frame with 122 rows and 3 variables:
\describe{
\item{tag}{OpenType four-character tag name}
\item{long_name}{Human-readable tag name}
\item{description}{Compact description of the tag}
}
}
\source{
\url{https://en.wikipedia.org/wiki/List_of_typographic_features}
}
\usage{
feature_dict
}
\description{
Compact data frame describing OpenType features.
}
\keyword{datasets}
