% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x-font-reg.R
\docType{data}
\name{gs_pkg}
\alias{gs_pkg}
\title{Goldman Sans & Goldman Sans Condensed Font Variants}
\format{
A list
}
\usage{
gs_pkg
}
\description{
On load hrbragg creates a custom Goldman Sans & Goldman Sans Condensed
font family variants and makes them available in the global environment.
}
\details{
The reconfigured family variant has the following features:

This object has the following names corresponding to individual,
customized font variants (ordered by font weight):

\strong{Goldman Sans} (\code{gs_rc})
\itemize{
\item \code{light}
\item \code{normal_italic}
\item \code{normal_light}
\item \code{medium}
\item \code{medium_italic}
\item \code{bold}
\item \code{bold_italic}
\item \code{heavy}
}

\strong{Goldman Sans Condensed} (\code{gsc_rc})
\itemize{
\item \code{normal}
\item \code{bold}
}
}
\keyword{datasets}
