% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrbragg-package.R
\docType{package}
\name{hrbragg}
\alias{hrbragg}
\alias{hrbragg-package}
\title{...}
\description{
A good description goes here otherwise CRAN checks fail.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/hrbragg}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/hrbragg/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
