% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-fonts.R
\name{install_clean_sans}
\alias{install_clean_sans}
\title{Install Clear Sans}
\usage{
install_clean_sans()
}
\description{
Clear Sans has minimized, unambiguous characters and slightly narrow proportions, making it
ideal for UI design. Its strong, recognizable forms avoid distracting ambiguity, making it
comfortable for reading short UI labels and long passages in both screen and print.
}
\note{
\href{https://github.com/intel/clear-sans/blob/main/LICENSE.txt}{Clear Sans Font License}
}
\examples{
install_clean_sans()
}
\references{
\href{https://github.com/intel/clear-sans}{Clean Sans}
}
