% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-fonts.R
\name{install_goldman_sans}
\alias{install_goldman_sans}
\alias{install_goldman_sans_condensed}
\title{Install Goldman Sans}
\usage{
install_goldman_sans()

install_goldman_sans_condensed()
}
\description{
Goldman Sans is a clean, modern typeface designed for dense data-rich
environments. From open letter shapes, enlarged x-height, and optical
balancing of individual letterforms, we designed Goldman Sans with
clarity in mind. Both normal and condensed versions are provided.
}
\note{
\href{https://design.gs.com/d/legal/goldman-sans-license}{Goldman Sans Restricted Font License}
}
\examples{
install_goldman_sans()
install_goldman_sans_condensed()
}
\references{
\href{https://design.gs.com/d/story/goldman-sans/}{Goldman Sans}
}
