% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-fonts.R
\name{install_inter}
\alias{install_inter}
\title{Install Inter}
\usage{
install_inter()
}
\description{
Inter is a typeface carefully crafted & designed for computer screens and
features a tall x-height to aid in readability of mixed-case and lower-case
text. Several OpenType features are provided as well, like contextual
alternates that adjusts punctuation depending on the shape of surrounding
glyphs, slashed zero for when you need to disambiguate "0" from "o",
tabular numbers, etc.
}
\note{
\href{https://choosealicense.com/licenses/ofl-1.1/}{SIL Open Font License 1.1}
}
\references{
\href{https://rsms.me/inter/}{Inter}
}
