% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-fonts.R
\name{install_roboto_condensed}
\alias{install_roboto_condensed}
\title{Install Roboto Condensed}
\usage{
install_roboto_condensed()
}
\description{
Roboto has a dual nature. It has a mechanical skeleton and the forms are
largely geometric. At the same time, the font features friendly and open
curves. While some grotesks distort their letterforms to force a rigid rhythm,
Roboto doesn’t compromise, allowing letters to be settled into their natural
width. This makes for a more natural reading rhythm more commonly found in
humanist and serif types.
}
\note{
\href{http://www.apache.org/licenses/LICENSE-2.0}{Apache License, Version 2.0.}
}
\examples{
install_roboto_condensed()
}
