% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{inter_pkg}
\alias{inter_pkg}
\title{Inter Font Variant}
\format{
A list
}
\usage{
inter_pkg
}
\description{
On load {hrbragg} creates a custom Inter font family variant
and makes it available in the global environment.
}
\details{
The reconfigured family variant has the following features:
\itemize{
\item \code{calt}: (\emph{Contextual Alternates}): Applies a second substitution feature
based on a match of a character pattern within a context of surrounding patterns
\item \code{case} (\emph{Case Sensitive Forms}): Replace characters, especially punctuation, with forms better suited for all-capital text, cf. titl
\item \code{dlig}/\code{ligatures} (\emph{Discretionary Ligatures}): Ligatures to be applied at the user's discretion
\item \code{kern} (\emph{Kerning}): Fine horizontal positioning of one glyph to the next, based on the shapes of the glyphs
\item \code{salt} (\emph{Stylistic Alternates}): \strong{DISABLED} — Either replaces with, or displays list of, stylistic alternatives for a character
\item \code{ss01} (\emph{Stylistic set 1: Alternate digits}): An alternate style of digits.
\item \code{tnum} (\emph{Tabular Figures}): Replaces numerals with glyphs of uniform width, often also lnum
\item \code{zero} (\emph{Slashed Zero}): \strong{DISABLED} — Replaces 0 figure with slashed 0
}

Inter online documentation has \href{https://rsms.me/inter/#features}{specific descriptions and exaples of these features}.

This object has the following names corresponding to individual,
customized font variants (ordered by font weight):
\itemize{
\item \code{ultralight}
\item \code{ultralight_italic}
\item \code{light}
\item \code{light_italic}
\item \code{normal_italic}
\item \code{normal}
\item \code{medium_italic}
\item \code{medium}
\item \code{semibold_italic}
\item \code{semibold}
\item \code{bold}
\item \code{bold_italic}
\item \code{ultrabold}
\item \code{ultrabold_italic}
\item \code{heavy}
\item \code{heavy_italic}
}
}
\keyword{datasets}
