% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-reference.R
\name{opentype_typographic_features}
\alias{opentype_typographic_features}
\title{OpenType Typographic Features}
\description{
Quick reference table describing the four-character feature codes. A
more complete quick reference can be found
\href{https://en.wikipedia.org/wiki/List_of_typographic_features}{on Wikipedia}
and a complete OpenType feature list specification can be found
\href{https://docs.microsoft.com/en-us/typography/opentype/spec/featurelist}{at Microsoft}.
}
\details{
\itemize{
\item \code{aalt} (\emph{Access All Alternates}): Special feature: used to present user with choice all alternate forms of the character
\item \code{abvf} (\emph{Above-base Forms}): Replaces the above-base part of a vowel sign. For Khmer and similar scripts.
\item \code{abvm} (\emph{Above-base Mark Positioning}): Positions a mark glyph above a base glyph.
\item \code{abvs} (\emph{Above-base Substitutions}): Ligates a consonant with an above-mark.
\item \code{afrc} (\emph{Alternative Fractions}): Converts figures separated by slash with alternative stacked fraction form
\item \code{akhn} (\emph{Akhand}): Hindi for unbreakable.  Ligates consonant+halant+consonant, usually only for k-ss and j-ny combinations.
\item \code{blwf} (\emph{Below-base Forms}): Replaces halant+consonant combination with a subscript form.
\item \code{blwm} (\emph{Below-base Mark Positioning}): Positions a mark glyph below a base glyph
\item \code{blws} (\emph{Below-base Substitutions}): Ligates a consonant with a below-mark.
\item \code{c2pc} (\emph{Capitals to Petite Caps}): Substitutes capital letters with petite caps
\item \code{c2sc} (\emph{Capitals to Small Caps}): Substitutes capital letters with small caps
\item \code{calt} (\emph{Contextual Alternates}): Applies a second substitution feature based on a match of a character pattern within a context of surrounding patterns
\item \code{case} (\emph{Case Sensitive Forms}): Replace characters, especially punctuation, with forms better suited for all-capital text, cf. titl
\item \code{ccmp} (\emph{Glyph Composition/Decomposition}): Either calls a ligature replacement on a sequence of characters or replaces a character with a sequence of glyphs. Provides logic that can for example effectively alter the order of input characters.
\item \code{cfar} (\emph{Conjunct Form After Ro}): Khmer
\item \code{cjct} (\emph{Conjunct Forms}): Ligates consonant+halant+consonant, indicates part of a conjunct.
\item \code{clig} (\emph{Contextual Ligatures}): Applies a second ligature feature based on a match of a character pattern within a context of surrounding patterns
\item \code{cpct} (\emph{Centered CJK Punctuation}): Positions punctuation marks vertically and horizontally
\item \code{cpsp} (\emph{Capital Spacing}): Adjusts spacing between letters in all-capitals text
\item \code{cswh} (\emph{Contextual Swash}): Converts letter to a swashed version based on characters around the letter
\item \code{curs} (\emph{Cursive Positioning}): Precise positioning of a letter's connection to an adjacent one
\item \verb{cv01–cv99} (\emph{Character Variant 1–99}): Multiple variants of a single character, which may not apply to many other characters, see references for voluminous documentation
\item \code{dist} (\emph{Distance}): Adjusts horizontal positioning between glyphs. (Always enabled, as opposed to 'kern'.)
\item \code{dlig} (\emph{Discretionary Ligatures}): Ligatures to be applied at the user's discretion
\item \code{dnom} (\emph{Denominator}): Converts to appropriate fraction denominator form, invoked by frac
\item \code{dtls} (\emph{Dotless Forms}):
\item \code{expt} (\emph{Expert Forms}): Typographic alternatives for some Japanese tetragrams
\item \code{falt} (\emph{Final Glyph on Line Alternates}): Replaces final glyph on the line with an alternate
\item \code{fin2} (\emph{Terminal Form #2}): Syriac
\item \code{fin3} (\emph{Terminal Form #3}): Syriac
\item \code{fina} (\emph{Terminal Forms}): Substitutes a special form of a letter at end of words (required by Arabic and Syriac)
\item \code{flac} (\emph{Flattened accent forms}):
\item \code{frac} (\emph{Fractions}): Converts figures separated by slash with diagonal fraction
\item \code{fwid} (\emph{Full Widths}): Substitutes proportionally spaced character with full-width versions (esp. for Latin letters within Chinese)
\item \code{half} (\emph{Half Form}): Replaces consonant+halant with a half form, indicating it is part of a conjunct.
\item \code{haln} (\emph{Halant Forms}): Replaces a glyph for final consonant+halant.
\item \code{halt} (\emph{Alternate Half Widths}): Re-positions full-width glyphs on half-width spaces
\item \code{hist} (\emph{Historical Forms}): Obsolete forms of characters to be applied at the user's discretion, cf. hlig
\item \code{hkna} (\emph{Horizontal Kana Alternates}): Alternate forms for horizontal kana text, e.g. ー for chōonpu instead of ｜, cf. vkna
\item \code{hlig} (\emph{Historical Ligatures}): Obsolete ligatures to be applied at the user's discretion
\item \code{hngl} (\emph{Hangul}): Transliterates Chinese-style characters with Korean Hangul
\item \code{hojo} (\emph{Hojo Kanji Forms}): Hojo alternates for Japanese tetragrams
\item \code{hwid} (\emph{Half Widths}): Substitutes uniformly-spaced characters with half-width version
\item \code{init} (\emph{Initial Forms}): Substitutes a special form of a letter occurring at the beginning of a word (required by Arabic and Syriac)
\item \code{isol} (\emph{Isolated Forms}): Substitutes a special form of a letter occurring outside a word (required by Arabic and Syriac)
\item \code{ital} (\emph{Italics}): Replaces letter with corresponding italic glyph
\item \code{jalt} (\emph{Justification Alternates}): User selectable wider and narrower alternates, used especially for justifying
\item \code{jp04} (\emph{JIS 04 Forms}): JIS 2004 alternates for Japanese tetragrams, not accessible per Unicode
\item \code{jp78} (\emph{JIS 78 Forms}): JIS C 6226-1978 alternates for Japanese tetragrams, not accessible per Unicode
\item \code{jp83} (\emph{JIS 83 Forms}): JIS X 0208-1983 alternates for Japanese tetragrams, not accessible per Unicode
\item \code{jp90} (\emph{JIS 90 Forms}): JIS X 0208-1990 alternates for Japanese tetragrams, not accessible per Unicode
\item \code{kern} (\emph{Kerning}): Fine horizontal positioning of one glyph to the next, based on the shapes of the glyphs
\item \code{lfbd} (\emph{Left Bounds}): Re-positions glyphs at end of line. Called by opbd.
\item \code{liga} (\emph{Standard Ligatures}): Replaces (by default) sequence of characters with a single ligature glyph
\item \code{ljmo} (\emph{Leading Jamo Forms}): Initial group of consonants for a synthesized Korean Hangul tetragram
\item \code{lnum} (\emph{Lining Figures}): Replaces numerals with glyphs meant to fit better in all-capitals text, often also tnum
\item \code{locl} (\emph{Localized Forms}): Substitutes character with the preferred form based on script language
\item \code{ltra} (\emph{Left-to-right glyph alternates}): Replaces characters with forms befitting left-to-right presentation (except mirrored forms)
\item \code{ltrm} (\emph{Left-to-right mirrored forms}): Replaces characters with possibly mirrored forms befitting left-to-right presentation
\item \code{mark} (\emph{Mark Positioning}): Fine positioning of a mark glyph to a base character
\item \code{med2} (\emph{Medial Form #2}): Syriac
\item \code{medi} (\emph{Medial Forms}): Substitutes a special form of letters between other letters in words (required by Arabic and Syriac)
\item \code{mgrk} (\emph{Mathematical Greek}): Replaces Greek characters with special forms for use in mathematics
\item \code{mkmk} (\emph{Mark-to-mark Positioning}): Fine positioning of a mark glyph to another mark character
\item \code{mset} (\emph{Mark Positioning via Substitution}): Used in Windows 95 for positioning of Arabic marks
\item \code{nalt} (\emph{Alternate Annotation Forms}): Provides user access to circled digits, inverse letters etc.
\item \code{nlck} (\emph{NLC Kanji Forms}): NLC alternates for Japanese tetragrams
\item \code{nukt} (\emph{Nukta Forms}): Replace consonant+nukta (dot mark) with single glyph.
\item \code{numr} (\emph{Numerator}): Converts to appropriate fraction numerator form, invoked by frac
\item \code{onum} (\emph{Oldstyle Figures}): Replaces numerals with cased old-style numerals, often also pnum
\item \code{opbd} (\emph{Optical Bounds}): Re-positions glyphs at beginning and end of line, for precise justification of text.
\item \code{ordn} (\emph{Ordinals}): Replaces characters with ordinal forms for use after numbers
\item \code{ornm} (\emph{Ornaments}): Decorative alternates for the bullet character •
\item \code{palt} (\emph{Proportional Alternates}): Re-positions otherwise monospace characters according to glyph width
\item \code{pcap} (\emph{Petite Caps}): Substitute lower-case letters with their petite caps analogs
\item \code{pkna} (\emph{Proportional Kana}): Kana for use alongside alphabets, without grid typography
\item \code{pnum} (\emph{Proportional Figures}): Replaces numerals with glyphs of proportional width, often also onum
\item \code{pref} (\emph{Pre-base Forms}): Replaces halant+consonant at the end of a consonant cluster with a glyph at the beginning. Khmer, Myanmar, Malayalam, Telugu
\item \code{pres} (\emph{Pre-base Substitutions}): Ligates consonant combinations.
\item \code{pstf} (\emph{Post-base Forms}): Substitutes final halant+consonant with special form.  Khmer and Gurmukhi, Malayalam
\item \code{psts} (\emph{Post-base Substitutions}): Ligates a final consonant+consonant.
\item \code{pwid} (\emph{Proportional Widths}): Replaces uniformly-spaced glyphs with proportional ones
\item \code{qwid} (\emph{Quarter Widths}): Replaces uniformly-spaced glyphs with quarter-width ones (punctuation etc.)
\item \code{rand} (\emph{Randomize}): Replaces character with random forms (meant to simulate handwriting)
\item \code{rclt} (\emph{Required Contextual Alternates}): Contextual alternates required for correct text display which differs from the default join for other letters, required especially important by Arabic
\item \code{rkrf} (\emph{Rakar Forms}): Replaces halant+ra with a rakar glyph, indicating it is part of a conjunct.
\item \code{rlig} (\emph{Required Ligatures}): Ligatures required for correct text display (any script, but in cursive)
\item \code{rphf} (\emph{Reph Form}): Replaces initial ra+halant with a final reph mark, indicating part of a conjunct.
\item \code{rtbd} (\emph{Right Bounds}): Re-positions glyphs at beginning of line. Called by opbd.
\item \code{rtla} (\emph{Right-to-left glyph alternates}): Replaces characters with forms befitting right-to-left presentation (except mirrored forms)
\item \code{rtlm} (\emph{Right-to-left mirrored forms}): Replaces characters with possibly mirrored forms befitting right-to-left presentation
\item \code{ruby} (\emph{Ruby Notation Forms}): Ruby characters, small print
\item \code{rvrn} (\emph{Required Variation Alternates}): Special variants of a single character, which need apply to specific font variation, required by variable fonts
\item \code{salt} (\emph{Stylistic Alternates}): Either replaces with, or displays list of, stylistic alternatives for a character
\item \code{sinf} (\emph{Scientific Inferiors}): as in "H2O", "SOx" or "YCbCr" (but using the same font weight and predefined position in contrast these plain HTML subs and sups)
\item \code{size} (\emph{Optical size}): Not a lookup: feature's table provides to applications information about the appearance and intent of the font, to aid in font selection.
\item \code{smcp} (\emph{Small Caps}): Substitutes lower-case letters with small caps versions
\item \code{smpl} (\emph{Simplified Forms}): Replaces sinograms with their simplified versions, may be language dependent
\item \code{ss01-ss20} (\emph{Stylistic Set 1 – 20}): Replaces character with one from a font-specific set of stylistic alternatives
\item \code{ssty} (\emph{Math script style alternates}):
\item \code{stch} (\emph{Stretching Glyph Decomposition}): Substitutes a special form of a stretchy glyph onto one or more letters (required by Syriac)
\item \code{subs} (\emph{Subscript}): Replaces character with subscript version, cf. numr
\item \code{sups} (\emph{Superscript}): Replaces character with superscript version, cf. dnom
\item \code{swsh} (\emph{Swash}): Either replaces character with or displays multiple swashed versions
\item \code{titl} (\emph{Titling Alternates}): Replaces characters with forms suited for large type, as in titles
\item \code{tjmo} (\emph{Trailing Jamo Forms}): Final group of consonants for a synthesized Korean Hangul tetragram
\item \code{tnam} (\emph{Traditional Name Forms}): Japanese alternates for proper names
\item \code{tnum} (\emph{Tabular Figures}): Replaces numerals with glyphs of uniform width, often also lnum
\item \code{trad} (\emph{Traditional Forms}): Replaces Chinese characters with their traditional versions
\item \code{twid} (\emph{Third Widths}): Substitutes uniformly-spaced character with a version of 1/3 width (punctuation, etc.)
\item \code{unic} (\emph{Unicase}): Replaces lowercase and uppercase letters with a set of single case glyphs
\item \code{valt} (\emph{Alternative Vertical Metrics}): Positions shorter characters to be centered vertically with full-height characters
\item \code{vatu} (\emph{Vattu Variants}): Replaces consonant+rakar combinations with a vattu variant ligature.
\item \code{vert} (\emph{Vertical Alternates}): A subset of vrt2: prefer the latter feature
\item \code{vhal} (\emph{Alternative Vertical Half Metrics}): Positions characters to be centered vertically with half-height characters
\item \code{vjmo} (\emph{Vowel Jamo Forms}): Medial group of vowels for a synthesized Korean Hangul tetragram
\item \code{vkna} (\emph{Vertical Kana}): Alternate Japanese kana forms for vertical text, e.g. ｜ for chōonpu instead of ー, cf. hkna
\item \code{vkrn} (\emph{Vertical Kerning}): Fine vertical positioning of characters based on shape
\item \code{vpal} (\emph{Proportional Alternate Vertical Metrics}): Re-positions glyphs vertically to be centered on proportional full-height characters
\item \code{vrt2} (\emph{Vertical Alternates and Rotation}): Replaces characters with forms suitable for vertical writing, possibly by rotating 90°
\item \code{vrtr} (\emph{Vertical Alternates for Rotation}): Replaces characters with forms suitable for vertical writing, possibly by shifting or shape
\item \code{zero} (\emph{Slashed Zero}): Replaces 0 figure with slashed 0
}
}
