% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview-variant.R
\name{preview_variant}
\alias{preview_variant}
\title{Preview numbers, kerning, and ligatures from font variants you create}
\usage{
preview_variant(
  font_variant,
  font_size = 5,
 
    test_label = "Lorem Ipsum dolor sit amet consectetur. => A\\u20dd\\n−0+1:2~3,456789\\n−9+8:7~6,543210"
)
}
\arguments{
\item{font_variant}{a font variant object created with \code{\link[=reconfigure_font]{reconfigure_font()}}}

\item{font_size}{passed to \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}}

\item{test_label}{a specially designed preview string to showcase some
specific font features}
}
\value{
A ggplot2 object
}
\description{
Use this function to preview the special features of fonts you create
including how well the default kerning ruls work, how tabular numbers
behave, and whether or not modern, fun ligatures are supported. Change
the \code{test_label} if you wish to use a custom preview string.
}
\examples{
reconfigure_font(
  family = "Trattatello",
  width = "normal",
  ligatures = "discretionary",
  calt = 1, tnum = 1, case = 1,
  dlig = 1, kern = 1,
  zero = 0, salt = 0
) -> trat

preview_variant(trat)

reconfigure_font(
  family = "Barlow",
  width = "normal",
  ligatures = "standard",
  tnum = 1, kern = 1
) -> barlow

preview_variant(barlow)
}
