% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconfigure-font.R
\name{reconfigure_font}
\alias{reconfigure_font}
\title{Create an complete, alternate font family with the same customized features}
\usage{
reconfigure_font(
  family,
  width = "normal",
  ligatures = NULL,
  letters = NULL,
  numbers = NULL,
  prefix = uuid::UUIDgenerate(),
  ...
)
}
\arguments{
\item{family}{font family name}

\item{width}{font width (defaults to "normal")}

\item{ligatures, letters, numbers}{See \code{\link[systemfonts:font_feature]{systemfonts::font_feature()}}}

\item{prefix}{string to use as a prefix to the names of the newly generated font family.
If not specified, a guaranteed unique identifier will be used.}

\item{...}{Extra four-letter font feature flags passed on to \code{\link[systemfonts:font_feature]{systemfonts::font_feature()}}. See
See \href{https://docs.microsoft.com/en-us/typography/opentype/spec/featurelist}{Microsoft} and
\href{https://en.wikipedia.org/wiki/List_of_typographic_features}{Wikipedia} for tag descriptions.}
}
\value{
list with names of all the customized font family members. A \code{family}
attribute is included which is the name of the family this variant is
based on.
}
\description{
Create an complete, alternate font family with the same customized features
}
\examples{
reconfigure_font(
  family = "Inter",
  width = "normal",
  ligatures = "discretionary",
  calt = 1, tnum = 1, case = 1,
  dlig = 1, ss01 = 1, kern = 1,
  zero = 0, salt = 0
) -> inter_1
}
