% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset-ggplot2-defaults.R
\name{reset_ggplot2_defaults}
\alias{reset_ggplot2_defaults}
\title{Restore all ggplot2 geom to default aesthetics}
\usage{
reset_ggplot2_defaults()
}
\description{
Restore all ggplot2 geom to default aesthetics
}
