% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrbragg-package.R
\docType{package}
\name{hrbragg}
\alias{hrbragg}
\alias{hrbragg-package}
\title{Typography-centric Themes, Theme Components, and Utilities for 'ggplot2' and 'ragg'.}
\description{
The 'ragg', 'systemfonts', and 'textshaping' packages make it
possible to create plot components and themes that make full use of
the rich typography features in modern fonts. Fonts, themes, and utilities
are provided to create 'ggplot2' plots intended for rendering on 'ragg'
graphics devices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/hrbragg}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/hrbragg/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
